+++ WrenServer/Source/Components/AIComponent.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "AIComponent.h"
 
-void AIComponent::Initialize(const unsigned int id, const long gameObjectId)
+void AIComponent::Initialize(const int id, const int gameObjectId)
 {
 	this->id = id;
 	this->gameObjectId = gameObjectId;
